Separate workspace scanning from reporting
